requirements.txt
pytest
RPi.GPIO == 0.5.4